Little bit safer deletion procedure

+++ src/main/scala/fi/liikennevirasto/digiroad2/util/JsonSerializer.scala
@@ -60,6 +60,8 @@ class JsonSerializer extends VVHSerializer {
   override def writeCache(file: File, objects: Seq[Object]): Boolean = {
 
     val tmpFile = File.createTempFile("tmp", "cached")
+    tmpFile.deleteOnExit()
+
     write(objects, new FileWriter(tmpFile))
 
     if (file.exists())
@@ -67,10 +69,15 @@ class JsonSerializer extends VVHSerializer {
 
     try {
       copy(Paths.get(tmpFile.getAbsolutePath), Paths.get(file.getAbsolutePath))
-      tmpFile.delete()
       return true
     } catch {
       case ex: Exception => logger.warn("Copy failed", ex)
+    } finally {
+      try {
+        if (tmpFile != null) tmpFile.delete()
+      } catch {
+        case ex: Exception => logger.info("Deleting tmp file failed", ex)
+      }
     }
     false
